Added functionality for validating password matching through react-hook-form
Removed the set up that was previously in place that manually checked that passwords matched, and set up the same functionality and more through react-hook-form. Set up state to track the password entered fiword entered first, then validated that the passwords must match on the verifyPassword input. Using validation for this through react-hook-form keeps validation consistent throughout this form since the rest of the fe rest of the form was using it as well, and allows users to see live error message updates for password matching