updated state that is stored upon login. Team name is now stored as well
added an additional axios call to grab team name, and to handle if user is not yet on a team.

still need to test functionality once a user is assigned, as of now no users in the database are assigned a team yet.

also removed REGISTER action, and replaced it with login, as they were doing the same thing.