removd toJSON function value that comes with error object in order to fix error state persisting bug
error state did not properly persist through refreshes. once stringified, the data that was stored to localStorage to persist data was different than the state stored to local state in redux. therefore, error messages worked before refresh, but if you refreshed when there was a current error on the screen, the data it would try to refrence upon refresh was different, and it couldn't find what it was looking for

deleted the toJSON function property that was on the error message that gets stored in state, since objects with a toJSON function property stringify differently. see documentation on developer.mozilla for .stringfy for more information on the cause