removed CarouselControl, setup to customize next/previous slide controls
removed CarouselControl component in order to customize the control settings for moving between slides.

passed down next/previous functions down as props to each of the metric slides.

in the WaterSlide component, passed down the next/previous functions again to the WaterQuestion component, and set up arrow buttons and attached the next/previous functions on as onClicks