leaderboard points now update accordingly when a user changes teams
added functionality to the existing updateUserTeam action in actions.js that is called when a user changes teams (in user-card.js). Now, when user changes teams, both the old team and the new team's points are adjusted on the back end accordingly.

still need to set up dispatch in reducer to update redux state for team points changes. However, this does not change affect the functionality for the app currently, just should make these changes for consistency