fixed bug where once new users are assigned to a team, existing points for that user are now added to designated team points
moved around axios calls, so that even if users don't have an assigned team, new team points can still be updated. Previously, that axios PUT request couldnt be hit as it was nested inside updating the old team axios call, which won't be hit in the case of a brand new user, as they had no previous team.